modif testunit

+++ src/main/java/com/inti/formation/service/CalculatriceService.java
@@ -2,9 +2,9 @@ package com.inti.formation.service;
 
 public class CalculatriceService {
 
-	public int addition(double a , double b) {
+	public int addition(int a , int b) {
 		
-		return (int) (a+b);
+		return a+b;
 		
 	}
